Refactors diff algorithm
