+++ src/vs/editor/browser/widget/diffEditorWidget2/diffEditorViewModel.ts
@@ -9,7 +9,7 @@ import { Disposable, toDisposable } from 'vs/base/common/lifecycle';
 import { IObservable, IReader, ISettableObservable, ITransaction, autorunWithStore, derived, observableSignal, observableSignalFromEvent, observableValue, transaction, waitForState } from 'vs/base/common/observable';
 import { readHotReloadableExport } from 'vs/editor/browser/widget/diffEditorWidget2/utils';
 import { ISerializedLineRange, LineRange } from 'vs/editor/common/core/lineRange';
-import { AdvancedLinesDiffComputer } from 'vs/editor/common/diff/advancedLinesDiffComputer';
+import { DefaultLinesDiffComputer } from 'vs/editor/common/diff/defaultLinesDiffComputer/defaultLinesDiffComputer';
 import { IDocumentDiff, IDocumentDiffProvider } from 'vs/editor/common/diff/documentDiffProvider';
 import { MovedText } from 'vs/editor/common/diff/linesDiffComputer';
 import { DetailedLineRangeMapping } from 'vs/editor/common/diff/rangeMapping';
@@ -175,7 +175,7 @@ export class DiffEditorViewModel extends Disposable implements IDiffEditorViewMo
 			debouncer.cancel();
 			contentChangedSignal.read(reader);
 			documentDiffProviderOptionChanged.read(reader);
-			readHotReloadableExport(AdvancedLinesDiffComputer, reader);
+			readHotReloadableExport(DefaultLinesDiffComputer, reader);
 
 			this._isDiffUpToDate.set(false, undefined);
