+++ src/vs/editor/common/diff/defaultLinesDiffComputer/algorithms/dynamicProgrammingDiffing.ts
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { OffsetRange } from 'vs/editor/common/core/offsetRange';
-import { IDiffAlgorithm, SequenceDiff, ISequence, ITimeout, InfiniteTimeout, DiffAlgorithmResult } from 'vs/editor/common/diff/algorithms/diffAlgorithm';
-import { Array2D } from 'vs/editor/common/diff/algorithms/utils';
+import { IDiffAlgorithm, SequenceDiff, ISequence, ITimeout, InfiniteTimeout, DiffAlgorithmResult } from 'vs/editor/common/diff/defaultLinesDiffComputer/algorithms/diffAlgorithm';
+import { Array2D } from 'vs/editor/common/diff/defaultLinesDiffComputer/utils';
 
 /**
  * A O(MN) diffing algorithm that supports a score function.
