+++ src/vs/editor/common/diff/defaultLinesDiffComputer/lineSequence.ts
@@ -1,0 +1,45 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { CharCode } from 'vs/base/common/charCode';
+import { OffsetRange } from 'vs/editor/common/core/offsetRange';
+import { ISequence } from 'vs/editor/common/diff/defaultLinesDiffComputer/algorithms/diffAlgorithm';
+
+export class LineSequence implements ISequence {
+	constructor(
+		private readonly trimmedHash: number[],
+		private readonly lines: string[]
+	) { }
+
+	getElement(offset: number): number {
+		return this.trimmedHash[offset];
+	}
+
+	get length(): number {
+		return this.trimmedHash.length;
+	}
+
+	getBoundaryScore(length: number): number {
+		const indentationBefore = length === 0 ? 0 : getIndentation(this.lines[length - 1]);
+		const indentationAfter = length === this.lines.length ? 0 : getIndentation(this.lines[length]);
+		return 1000 - (indentationBefore + indentationAfter);
+	}
+
+	getText(range: OffsetRange): string {
+		return this.lines.slice(range.start, range.endExclusive).join('\n');
+	}
+
+	isStronglyEqual(offset1: number, offset2: number): boolean {
+		return this.lines[offset1] === this.lines[offset2];
+	}
+}
+
+function getIndentation(str: string): number {
+	let i = 0;
+	while (i < str.length && (str.charCodeAt(i) === CharCode.Space || str.charCodeAt(i) === CharCode.Tab)) {
+		i++;
+	}
+	return i;
+}
