+++ src/vs/editor/test/node/diffing/advancedLinesDiffComputer.test.ts
@@ -6,7 +6,7 @@
 import * as assert from 'assert';
 import { Range } from 'vs/editor/common/core/range';
 import { RangeMapping } from 'vs/editor/common/diff/rangeMapping';
-import { LinesSliceCharSequence, getLineRangeMapping } from 'vs/editor/common/diff/advancedLinesDiffComputer';
+import { LinesSliceCharSequence, getLineRangeMapping } from 'vs/editor/common/diff/defaultLinesDiffComputer/defaultLinesDiffComputer';
 import { OffsetRange } from 'vs/editor/common/core/offsetRange';
 
 suite('lineRangeMapping', () => {
